requirements.txt
Django==1.7
MySQL-python==1.2.5
lxml==3.4.0
python-dateutil==2.2
six==1.8.0
vanity==2.0.5
wsgiref==0.1.2
